[Refactor] Move lezer to src/js folder
