+++ tsconfig.json
@@ -8,5 +8,8 @@
       "sourceMap": true,
       "outDir": "dist/"
     },
-    "lib": ["es2015"]
+    "lib": ["es2015"],
+    "include": [
+      "./src/**/*"
+    ]
   }
